requirements.txt
Django
djangorestframework
psycopg2-binary
psycopg2
flake8
pytest
pytest-django
pytest-factoryboy
pytest-mock
ipdb
python-decouple
python-memcached
requests